+++ source/ftrack_connect_nuke/ui/delegate.py
@@ -1,14 +1,15 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014 ftrack
 
-import functools
 import FnAssetAPI
-from FnAssetAPI.ui.toolkit import QtGui
 from ftrack_connect_foundry.ui import delegate
 from ftrack_connect_foundry.ui.tasks_view import TasksView as _TasksView
 from ftrack_connect_foundry.ui.info_view import (
     WorkingTaskInfoView as _WorkingTaskInfoView, InfoView as _InfoView
 )
+
+import ftrack_connect.ui.theme
+
 
 class Delegate(delegate.Delegate):
     def __init__(self, bridge):
@@ -21,15 +22,16 @@ class Delegate(delegate.Delegate):
         import nuke
         import legacy
         from nukescripts import panels
+
+        from ftrack_connect_nuke.ui.widget.crew import NukeCrew
         from ftrack_connect_nuke.connector import Connector
 
         from ftrack_connect_nuke.ui.widget.publish_gizmo import GizmoPublisherDialog
 
-        from ftrack_connect_nuke.ui.widget.publish_script import ScriptPublisherDialog
-        from ftrack_connect_nuke.ui.widget.load_script import ScriptOpenerDialog
         Connector.registerAssets()
 
-        # wrappers for initializing the widgets with the correct connector object
+        # wrappers for initializing the widgets with
+        # the correct connector object
         def wrapImportAssetDialog(*args, **kwargs):
             from ftrack_connect.ui.widget.import_asset import FtrackImportAssetDialog
             return FtrackImportAssetDialog(connector=Connector)
@@ -37,7 +39,6 @@ class Delegate(delegate.Delegate):
         def wrapAssetManagerDialog(*args, **kwargs):
             from ftrack_connect.ui.widget.asset_manager import FtrackAssetManagerDialog
             return FtrackAssetManagerDialog(connector=Connector)
-
 
         # Populate the ui
         nukeMenu = nuke.menu("Nuke")
@@ -104,16 +105,24 @@ class Delegate(delegate.Delegate):
                 )
             )
 
+        ftrackMenu.addSeparator()
+
+        # Create the crew dialog entry in the menu
+        panels.registerWidgetAsPanel(
+            'ftrack_connect_nuke.ui.widget.crew.NukeCrew',
+            'Crew',
+            'widget.Crew'
+        )
+        ftrackMenu.addCommand(
+            'Crew',
+            'pane = nuke.getPaneFor("Properties.1");'
+            'panel = nukescripts.restorePanel("widget.Crew");'
+            'panel.addToPane(pane)'
+        )
 
         # Add new entries in the ftrack menu.
         ftrackMenu.addSeparator()
         ftrackMenu.addCommand('Publish gizmo', GizmoPublisherDialog)
-
-        # The new load and publish script dialog's are waiting for some style
-        # fixes.
-        if False:
-            ftrackMenu.addCommand('Publish script', ScriptPublisherDialog)
-            ftrackMenu.addCommand('Load script', ScriptOpenerDialog)
 
         # Add ftrack publish node
         toolbar = nuke.toolbar("Nodes")
@@ -136,3 +145,7 @@ class Delegate(delegate.Delegate):
 
         if host and host.getIdentifier() == 'uk.co.foundry.nuke':
             self.populate_ftrack()
+
+            # Set font on QApplication once UI is created.
+            # We do this once since it takes some time to apply the font.
+            ftrack_connect.ui.theme.applyFont()